Muokattu komentoa

+++ Moduuli 4/Python moduuli 4 teht 2.py
@@ -1,6 +1,6 @@
-tuuma = int(input('Anna tuumat: '))
+tuuma = float(input('Anna tuumat: '))
 sentti = 2.54
 while tuuma>0:
     print(f'tuumat sentteinä: {tuuma*sentti}')
-    tuuma = int(input('Anna tuumat: '))
+    tuuma = float(input('Anna tuumat: '))
 print("lopetan toimintani")
